pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-codevf"
version = "0.1.0"
description = "LangChain Tool integration for CodeVF"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "langchain-core>=0.1.0",
  "pydantic>=2.0",
  "codevf-sdk>=0.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-asyncio>=0.23",
  "python-dotenv>=1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
